refactor(core): group status-bar extensions under statusbar/
Move panel-toggles, settings-button, and updates into
extensions-core/src/statusbar/ to keep the package root focused on
domain extensions. These three are the standalone status-bar items —
the ones without their own domain/panel (unlike themes/workspaces,
whose status items live with their service).

No behavior change: all imports were folder-local, so only the barrel's
import paths move.
