+++ packages/extensions-core/src/index.ts
@@ -15,6 +15,6 @@ export { extension as themes } from "./themes";
 export { extension as keybindings } from "./keybindings";
 export { extension as about } from "./about";
 export { extension as extensions } from "./extensions";
-export { extension as panelToggles } from "./panel-toggles";
-export { extension as settingsButton } from "./settings-button";
-export { extension as updates } from "./updates";
+export { extension as panelToggles } from "./statusbar/panel-toggles";
+export { extension as settingsButton } from "./statusbar/settings-button";
+export { extension as updates } from "./statusbar/updates";
