+++ packages/extensions-core/src/statusbar/panel-toggles/index.tsx
