+++ packages/extensions-core/src/statusbar/panel-toggles/panel-toggles.css
