+++ packages/extensions-core/src/statusbar/settings-button/settings-button.css
