+++ packages/extensions-core/src/statusbar/updates/UpdatePrompt.tsx
