+++ packages/extensions-core/src/statusbar/updates/index.tsx
