+++ packages/extensions-core/src/statusbar/updates/model.test.ts
