+++ packages/extensions-core/src/statusbar/updates/model.ts
