+++ packages/extensions-core/src/statusbar/updates/updates.css
